pyproject.toml
[tool.poetry]
name = "application-manager"
version = "0.0"
description = "Description of your tool"
authors = ["Marco Simoni marco.simoni@iit.cnr.it"]
repository = "https://github.com/sifis-home/Application-Manager"
license = "MIT"
classifiers = [
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
]
include = ["README.md", "LICENSE.md", "application-manager/data/*"]
packages = [
    { include = "application-manager" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
docker = "6.0.1"
websocket-client = "1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pre-commit = "^3.0.4"
black = "^23.1.0"
isort = "^5.12.0"
ruff = "^0.0.63"
coverage = {extras = ["toml"], version = "^7.1.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
disallow_untyped_decorators = false

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.pylint.master]
jobs = 0

[tool.pylint.basic]
good-names = [
    "ex","_","w","h","i","j","x","y","x1","x2","y1","y2"
]

[tool.pylint.format]
max-line-length = 79

[tool.pylint.messages_control]
reports = false
disable = [
    "import-error",
    "missing-docstring",
    "unused-argument",
    "duplicate-code",
    "too-many-return-statements",
    "too-many-arguments",
    "too-many-locals",
    "too-many-instance-attributes",
    "too-many-public-methods",
    "too-few-public-methods",
    "no-self-use",
]

[tool.pylint.miscellaneous]
notes = ["FIXME", "TODO"]

[tool.ruff]
line-length = 79
select = [
    "F401",
    "F403",
]

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 60
